Spring security added and JWT support with Login and Signup endpoints working

+++ 02-backend/spring-boot-ecommerce/src/main/java/com/thiaghoul/ecommerce/entities/Country.java
@@ -17,7 +17,7 @@ public class Country{
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private int id;
+    private Short id;
 
     @Column(name = "code")
     private String code;
